backend/requirements.txt
fastapi==0.133.1            # API framework
uvicorn[standard]==0.32.1   # Development server for FastAPI
python-dotenv==1.0.1        # Load .env environment variables
supabase==2.28.0            # Supabase Python client
requests==2.32.3            # HTTP library
pandas
numpy